add immutable return in get method refactor naming
