--- src/AlphabetsForCipher/CyrillicAlphabetForCipher.java
@@ -1,54 +1,0 @@
-package AlphabetsForCipher;
-
-import java.util.ArrayList;
-
-//TODO Think about adding decorator
-public class CyrillicAlphabetForCipher extends AlphabetForCipher {
-    private static final CyrillicAlphabetForCipher CYRILLIC_ALPHABET_FOR_CIPHER = new CyrillicAlphabetForCipher();
-    private static final int START_SYMBOLS_CODE = 32;
-    private static final int END_SYMBOLS_CODE = 47;
-    private static final int YO_IN_LOWER_CODE = 1105;
-    private static final int YO_IN_UPPER_CODE = 1025;
-    private static final int START_CYRILLIC_ALPHABET_CODE = 1040;
-    private static final int END_CYRILLIC_ALPHABET_CODE = 1103;
-
-    private final ArrayList<Integer> SYMBOLS = new ArrayList<>();
-    private final ArrayList<Integer> CYRILLIC_ALPHABET_CODES = new ArrayList<>();
-    private final ArrayList<Integer> CYRILLIC_ALPHABET_CODES_AND_SYMBOLS = new ArrayList<>();
-    public ArrayList<Integer> cyrillicAlphabetCodesAndSymbols = new ArrayList<>();
-
-
-    private CyrillicAlphabetForCipher(){
-        setSymbolsCodes();
-        setCyrillicAlphabetCodes();
-        CYRILLIC_ALPHABET_CODES_AND_SYMBOLS.addAll(CYRILLIC_ALPHABET_CODES);
-        CYRILLIC_ALPHABET_CODES_AND_SYMBOLS.addAll(SYMBOLS);
-        cyrillicAlphabetCodesAndSymbols.addAll(CYRILLIC_ALPHABET_CODES_AND_SYMBOLS);
-    }
-    //Create one object instance of cyrillic alphabet
-    public static CyrillicAlphabetForCipher getInstance(){
-        return CYRILLIC_ALPHABET_FOR_CIPHER;
-    }
-
-    //Merge cyrillic alphabet codes with general symbols codes and return new object instance
-    public ArrayList<Integer> getAlphabetCodesAndSymbols(){
-        return cyrillicAlphabetCodesAndSymbols;
-    }
-
-    //Fill in cyrillic alphabet codes
-    private void setCyrillicAlphabetCodes(){
-        CYRILLIC_ALPHABET_CODES.add(YO_IN_UPPER_CODE);
-        CYRILLIC_ALPHABET_CODES.add(YO_IN_LOWER_CODE);
-        for (int i = START_CYRILLIC_ALPHABET_CODE, j = 0; i <= END_CYRILLIC_ALPHABET_CODE; i++, j++) {
-            CYRILLIC_ALPHABET_CODES.add(i);
-        }
-    }
-
-    //Fill in general symbols codes
-    private void setSymbolsCodes(){
-        for (int i = START_SYMBOLS_CODE; i <= END_SYMBOLS_CODE; i++) {
-            SYMBOLS.add(i);
-        }
-    }
-
-}
